Major organization overhaul, again
A lot of common files are now symlinked. I am a bit squiffy about what
git is going to do to my symlinks, but I am hopeful. Next step: Build
script to take this elaborate version, which is sort of like the
"instructor's" version with all the answers, and convert it into a
"student" version that will be referenced in the book.